added mc event weight and pile up weights back in the output
git-svn-id: svn+ssh://svn.cern.ch/reps/penn/bjackson/Susy2.0/trunk@19930 4525493e-7705-40b1-a816-d608a930855b

+++ utils/CutFlowDump/include/CutFlowDump.h
@@ -36,7 +36,7 @@ public :
    ULong64_t m_event_desc;
    ULong64_t m_sr_helper;
 
-   float     m_mc_weight;
+   double    m_mc_weight;
    double    m_pile_up_weight;
    double    m_trigger_weight;
    double    m_lepton_weight;
